requirements.txt
annotated-types==0.5.0
anyio==4.0.0
black==23.7.0
boto3==1.28.39
botocore==1.31.39
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
flake8==6.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
